requirements.txt
librosa
Pillow
numpy
matplotlib
pre-commit
scikit-learn
